requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
eventlet==0.39.1
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SocketIO==5.5.1
greenlet==3.2.1
h11==0.16.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
pyserial==3.5
pytest==8.3.5
python-engineio==4.12.0
python-socketio==5.13.0
simple-websocket==1.1.0
tomli==2.2.1
Werkzeug==3.1.3
wsproto==1.2.0
